chore: create separate components folder for all components
