+++ packages/react-tinacms-editor/src/components/Editor/styles/ProseMirror.tsx
